requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
batchgenerators==0.25
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
Cython==0.29.37
decorator==5.1.1
dicom2nifti==2.5.0
einops==0.8.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
fasteners==0.15
filelock==3.16.1
fonttools==4.54.1
free-mujoco-py==2.1.6
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
glfw==1.12.0
hiddenlayer==0.3
huggingface-hub==0.26.1
hydra-core==1.3.2
idna==3.10
imageio==2.36.0
importlib_resources==6.4.5
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
lazy_loader==0.4
lightning-utilities==0.11.9
linecache2==1.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
MedPy==0.5.2
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
nibabel==5.3.2
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydicom==3.0.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-gdcm==3.0.24.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
SimpleITK==2.4.0
six==1.16.0
stack-data==0.6.3
surface-distance==0.1
sympy==1.13.3
tensorboardX==2.6.2.2
tensorly==0.9.0
termcolor==2.5.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==1.0.9
torch @ file:///mnt/c/Nick/pytorch/whl/3.10_linux/torch-2.1.2%2Bcu118-cp310-cp310-linux_x86_64.whl
torchinfo==1.8.0
torchmetrics==1.6.0
torchprofile==0.0.4
torchsummary==1.5.1
torchvision @ file:///mnt/c/Nick/pytorch/whl/3.10_linux/torchvision-0.16.2%2Bcu118-cp310-cp310-linux_x86_64.whl
tornado==6.4.1
tqdm==4.66.5
traceback2==1.4.0
traitlets==5.14.3
triton==2.1.0
typing_extensions==4.12.2
tzdata==2024.2
unittest2==1.1.0
urllib3==2.2.3
visdom==0.2.4
wcwidth==0.2.13
websocket-client==1.8.0
yacs==0.1.8
yarl==1.16.0
